requirements.txt
mediapipe~=0.8.9.1
opencv-python~=4.5.5.64
numpy~=1.20.3
tensorflow-gpu~=2.8.0